chore(util): moves Util
